pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "archi-bot"
version = "0.1.0"
description = "A Discord Bot for managing Archipelago Randomizers"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "beautifulsoup4>=4.13",
    "hikari-arc>=2.0.0",
    "hikari-miru>=4.2.0",
    "hikari[speedups]>=2.2.0",
    "httpx>=0.28.1",
    "orjson>=3.10.16",
    "pydantic>=2.11.1",
    "python-dotenv>=1.1.0",
    "ruff>=0.11.2",
    "sqlmodel>=0.0.24",
    "uvloop>=0.21.0",
    "websockets>=15.0.1",
]

[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.setuptools]
packages = ["archi_bot"]

[tool.ruff]
preview = true

[tool.ruff.lint]
select = [
    "A",
    "ANN",
    "ARG",
    "ASYNC",
    "B",
    "BLE",
    "C4",
    "COM818",
    "D",
    "DTZ",
    "EM",
    "F",
    "FLY",
    "FURB",
    "ICN",
    "INP",
    "ISC",
    "N",
    "PERF",
    "PTH",
    "Q",
    "RET",
    "RUF",
    "S",
    "SIM",
    "SLF",
    "TC",
    "TID",
    "TRY",
    "UP",
]

[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = ["arc.inject"]

[tool.ruff.lint.flake8-builtins]
ignorelist = ["open"]

[tool.ruff.lint.pydocstyle]
convention = "google"
ignore-decorators = ["arc.loader", "arc.unloader"]
